training_data/requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
tensorflow-cpu>=2.10.0
scikit-learn>=1.1.0
matplotlib>=3.5.0 